refactor(registry): rename sample files to start with "sample-"
