+++ registry/src/preact/sample/sample-user-data.ts
