+++ registry/src/react/sample/sample-user-data.ts
