+++ registry/src/solid/sample/sample-tag-data.ts
