+++ registry/src/solid/sample/sample-user-data.ts
