+++ registry/src/svelte/examples/user-menu-dynamic/use-user-query.svelte.ts
@@ -1,7 +1,7 @@
 import {
   queryUsers,
   type User,
-} from '../../sample/query-users'
+} from '../../sample/sample-query-users'
 
 export function useUserQuery(
   getQuery: () => string,
