+++ registry/src/svelte/sample/sample-tag-data.ts
