+++ registry/src/svelte/sample/sample-user-data.ts
