+++ registry/src/vue/examples/user-menu/editor.vue
@@ -5,8 +5,8 @@ import 'prosekit/basic/typography.css'
 import { createEditor } from 'prosekit/core'
 import { ProseKit } from 'prosekit/vue'
 
-import { tags } from '../../sample/tag-data'
-import { users } from '../../sample/user-data'
+import { tags } from '../../sample/sample-tag-data'
+import { users } from '../../sample/sample-user-data'
 import { TagMenu } from '../../ui/tag-menu'
 import { UserMenu } from '../../ui/user-menu'
