+++ registry/src/vue/sample/sample-user-data.ts
